feat: Add SGBD Manager and OJDBC Driver
